Abstract DomainTranslator in SPI
